Fix: GetPixel() Changelog: GetPixel seemed not to like to work every time, when the function type was declared as 'long' and not 'int'.

+++ src/ffscript.h
@@ -119,7 +119,7 @@ void SetFFEngineFlag(int flag, bool v);
 void SetItemMessagePlayed(int itm);
 bool GetItemMessagePlayed(int itm);
 
-long do_getpixel();
+int do_getpixel();
 
 void do_getDMapData_dmapname(const bool v);
 void do_setDMapData_dmapname(const bool v);
